% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{benzing1977}
\alias{benzing1977}
\title{Halle publishers of the 16th and 17th centuries}
\format{
An object of class \code{"data.frame"}
}
\source{
\url{https://github.com/haldru/Benzing1977}
}
\usage{
data(benzing1977)
}
\description{
GND identifiers and preferred names of persons and corporate bodies involved in the Halle publishing business.
}
\keyword{datasets}
