% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{benzing1982}
\alias{benzing1982}
\title{Halle book printers of the 16th and 17th centuries}
\format{
An object of class \code{"data.frame"}
}
\source{
\url{https://github.com/haldru/Benzing1982}
}
\usage{
data(benzing1982)
}
\description{
GND identifiers and preferred names of persons involved in the Halle printing business.
}
\keyword{datasets}
