% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{paisey1988}
\alias{paisey1988}
\title{Halle printers, booksellers and publishers: 1701–1750}
\format{
An object of class \code{"data.frame"}
}
\source{
\url{https://github.com/haldru/Paisey1988}
}
\usage{
data(paisey1988)
}
\description{
GND identifiers and preferred names of persons and corporate bodies involved in the Halle printing business.
}
\keyword{datasets}
