% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{paisey1989}
\alias{paisey1989}
\title{Halle printers, booksellers and publishers of the 17th century}
\format{
An object of class \code{"data.frame"}
}
\source{
\url{https://github.com/haldru/Paisey1989}
}
\usage{
data(paisey1989)
}
\description{
GND identifiers and preferred names of persons involved in the Halle printing business.
}
\keyword{datasets}
