% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reske2015}
\alias{reske2015}
\title{Halle printers of the 16th and 17th centuries}
\format{
An object of class \code{"data.frame"}
}
\source{
\url{https://github.com/haldru/Reske2015}
}
\usage{
data(reske2015)
}
\description{
GND identifiers and preferred names of persons involved in the Halle printing business.
}
\keyword{datasets}
